+++ app/cdap/components/GridTable/components/FunctionNameToggle/index.tsx
@@ -15,69 +15,23 @@
  */
 
 import React from 'react';
-import { styled } from '@material-ui/core';
-import Switch, { SwitchProps } from '@material-ui/core/Switch';
 import { Typography, Box } from '@material-ui/core';
 import { useStyles } from 'components/GridTable/components/FunctionNameToggle/styles';
-import grey from '@material-ui/core/colors/grey';
 import T from 'i18n-react';
 import { IFunctionNameToggleProps } from 'components/GridTable/components/FunctionNameToggle/types';
-
-const StyledSwitch = styled(Switch)(({ theme }) => ({
-  width: 28,
-  height: 16,
-  padding: 0,
-  display: 'flex',
-  '&:active': {
-    '& .MuiSwitch-thumb': {
-      width: 15,
-    },
-    '& .MuiSwitch-switchBase.Mui-checked': {
-      transform: 'translateX(9px)',
-    },
-  },
-  '& .MuiSwitch-switchBase': {
-    padding: 2,
-    color: grey[600],
-    '&.Mui-checked': {
-      transform: 'translateX(12px)',
-      color: grey[600],
-      '& + .MuiSwitch-track': {
-        opacity: 1,
-        backgroundColor: '#ffffff',
-        border: `1px solid ${grey[600]}`,
-      },
-    },
-  },
-  '& .MuiSwitch-thumb': {
-    boxShadow: '0 2px 4px 0 #00230b33',
-    width: 12,
-    height: 12,
-    borderRadius: 6,
-    transition: theme.transitions.create(['width'], {
-      duration: 200,
-    }),
-  },
-  '& .MuiSwitch-track': {
-    borderRadius: 8,
-    opacity: 1,
-    backgroundColor: '#ffffff',
-    border: `1px solid ${grey[600]}`,
-    boxSizing: 'border-box',
-  },
-}));
+import SwitchInputComponent from 'components/common/Switch';
 
 export default function({ setIsShowName, isShowNames }: IFunctionNameToggleProps) {
   const classes = useStyles();
   return (
     <Box className={classes.functionWrapper}>
-      <Typography className={classes.typoClass} component='div'>
+      <Typography className={classes.typoClass} component="div">
         {T.translate('features.WranglerNewUI.GridPage.toolbarIcons.labels.toggleDescription')}
         &nbsp;
       </Typography>
-      <StyledSwitch
-        onClick={() => setIsShowName(!isShowNames)}
-        checked={isShowNames}
+      <SwitchInputComponent
+        setIsShowName={setIsShowName}
+        isShowNames={isShowNames}
         inputProps={{
           'aria-label': T.translate(
             'features.WranglerNewUI.GridPage.gridHeader.ariaLabels.functionsName'
